fix: remove value ambiguity
We accept strings as values to publish, but the actual datatype is
Uint8Array.  Likewise these Uint8Arrays get turned into strings
when passed back, but in JavaScript this can be a lossy translation
as bytes values above 127 are interpreted as multi-byte UTF8 code points.

This PR removes strings as acceptible input types and only accepts
Uint8Arrays.

BREAKING CHANGE: strings are no longer accepted as valid values to publish