fix: parse peer id from message correctly
Peer IDs are sent as a buffer containing the utf bytes for `/ipns/`
then the bytes that make up the peer id.

Our tests were using a buffer that contained the utf bytes for `/ipns/`
followed by the utf bytes for the base encoded peer id, which is wrong.

I think this only worked previously because the validator passed the
bytes to `PeerId.createFromBytes` which passes them to the `cids` `CID`
class constructor which is a lot more forgiving in terms of input than
the multiformats `CID` class.