feat!: opt-in V2-only records, IPIP-428 verification (#234)
Adds a `v1Compatible` option when creating IPNS records that will cause the V1 signature
to be added to the record.  This option defaults to true, in the future it will be changed to
false.

The value types have also been updated to make it harder to create invalid records.  It now
accepts only CIDs, PeerIds, or arbitrary path strings prefixed with `"/"`.

BREAKING CHANGE: all /ipns/* keys are now encoded as base36 encoded CIDv1 libp2p-cid

Closes #217

---------

Co-authored-by: Marcin Rataj <lidel@lidel.org>
Co-authored-by: Alex Potsides <alex@achingbrain.net>